requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.17.0
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
prophet==1.1.4
statsmodels==0.14.0
jupyter==1.0.0
openpyxl==3.1.2
fpdf2==2.7.6
sqlalchemy==2.0.23
redis==5.0.1
websockets==11.0.3
pytest==7.4.3
black==23.9.1
flake8==6.1.0
faker==19.6.2
